data modeling refactor
refactor data modeling
